requirements.txt
django==4.0
pillow==9.0.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
openpyxl==3.0.9
